Código refatorado para uma melhor leitura. Reutilizando a classe conexão nos acessos ao database.
