--- src/main/java/com/estacionamento/ControladorLogin
@@ -1,24 +1,0 @@
-package com.estacionamento;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.TextField;
-
-public class ControladorLogin {
-
-    @FXML
-    private Button botaoLogin;
-
-    @FXML
-    private TextField campoSenha;
-
-    @FXML
-    private TextField campoUsuario;
-
-    @FXML
-    void confirmarDados(ActionEvent event) {
-
-    }
-
-}
