+++ src/main/java/com/estacionamento/Mainlayout.java
@@ -1,10 +1,10 @@
 package com.estacionamento;
 
 import java.net.URL;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.ResourceBundle;
-
-import com.estacionamento.DAO.Conexao;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -13,6 +13,8 @@ import javafx.fxml.Initializable;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
+
+import com.estacionamento.DAO.Conexao;
 
 public class Mainlayout implements Initializable {
 
@@ -45,10 +47,44 @@ public class Mainlayout implements Initializable {
 
         // Chamar o método listar() da classe Conexao para obter os dados do banco de
         // dados
-        ArrayList<Estacionamento> listaEstacionamentos = Conexao.listar();
+        ArrayList<Estacionamento> listaEstacionamentos = listar();
         dados.addAll(listaEstacionamentos);
 
         tabela_estaciona.setItems(dados);
+
+    }
+
+    /*
+     * (Connection connection = DriverManager
+     * .getConnection("jdbc:sqlite:D:\\java\\DB\\estacionamento.db"))
+     */
+    public static ArrayList<Estacionamento> listar() {
+        try {
+            Conexao connection = new Conexao();
+            ArrayList<Estacionamento> response = new ArrayList<Estacionamento>();
+            PreparedStatement prepared_statement = connection.conectarBD()
+                    .prepareStatement("select * from estacionamento");
+            ResultSet result_set = prepared_statement.executeQuery();
+
+            while (result_set.next()) {
+                Estacionamento estacionamento = new Estacionamento(
+                        result_set.getString("nome"),
+                        result_set.getString("carro"),
+                        result_set.getString("placa"),
+                        result_set.getString("vaga"));
+
+                response.add(estacionamento);
+            }
+            result_set.close();
+            prepared_statement.close();
+
+            return response;
+        } catch (Exception e) {
+
+            System.out.println(e.getMessage());
+            return null;
+
+        }
 
     }
